+++ Makefile
@@ -1,5 +1,5 @@
 CC = cc
-CFLAGS = #-Wall -Wextra -Werror -ggdb3 -g #-fsanitize=address
+CFLAGS = -Wall -Wextra -Werror #-fsanitize=address
 LIBFT_DIR = libft
 LIBFT = $(LIBFT_DIR)/libft.a
 
@@ -7,7 +7,7 @@ SRC = execution/builtins/do_cd.c execution/builtins/do_echo.c execution/builtins
 	  execution/builtins/do_pwd.c execution/builtins/do_unset.c \
       execution/pipe_ex_helpers.c  execution/pipe_ex_helpers2.c execution/utils.c execution/utils2.c execution/dup_utils.c execution/redirections.c execution/pipe_executor.c execution/main.c execution/main_helpers.c\
       lexer/error.c lexer/expand.c lexer/lexer_help.c lexer/main.c lexer/parce.c lexer/t_cmd_helpers.c lexer/token_help.c lexer/tokenizing.c \
-      signals.c lexer/exit_status.c lexer/herdoc_parser.c execution/herdoc_executter.c lexer/createnv.c lexer/wild_card.c lexer/print_lexer.c
+      signals.c lexer/exit_status.c lexer/herdoc_parser.c execution/herdoc_executter.c lexer/createnv.c lexer/wild_card.c lexer/print_lexer.c lexer/wild_card_help.c
 
 OBJ = $(SRC:%.c=%.o)
 NAME = minishell
