+++ lexer/parce.h
@@ -6,7 +6,7 @@
 /*   By: hfakou <hfakou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/13 17:34:33 by hfakou            #+#    #+#             */
-/*   Updated: 2025/08/12 23:03:55 by hfakou           ###   ########.fr       */
+/*   Updated: 2025/08/13 06:47:58 by hfakou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include "../signals.h"
 # include "lexer.h"
 # include <dirent.h>
-#include <stdbool.h>
+# include <stdbool.h>
 
 typedef enum s_redir_type
 {
@@ -95,5 +95,10 @@ char				*ft_strstr(char *str, char *sub, size_t subsz);
 int					wds(t_token *tok);
 void				asterisk_or_args(char *arg, t_cmd *cmd);
 
-bool	next_joined_word_is_pattern(t_lexer *lexer, t_env *env);
+bool				next_joined_word_is_pattern(t_lexer *lexer);
+int					asterisk_in_filename(char *target, t_cmd *cmd,
+						t_token *tok);
+char				*get_single_file_or_null(char *patern);
+char				*join_current_dir_redi(char *patern);
+
 #endif
