+++ lexer/wild_card_help.c
@@ -1,8 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   wild_card_help.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hfakou <hfakou@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/08/13 06:51:27 by hfakou            #+#    #+#             */
+/*   Updated: 2025/08/13 07:06:01 by hfakou           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "parce.h"
 
-int asterisk_in_filename(char *target, t_cmd *cmd, t_token *tok)
+void	asterisk_or_args(char *arg, t_cmd *cmd)
 {
-	char *ast_tar;
+	if (arg && ft_strchr(arg, '*'))
+	{
+		if (!join_current_dir(cmd, arg))
+			add_to_argv(cmd, arg);
+		else
+			free(arg);
+	}
+	else if (arg)
+		add_to_argv(cmd, arg);
+}
+
+int	asterisk_in_filename(char *target, t_cmd *cmd, t_token *tok)
+{
+	char	*ast_tar;
 
 	ast_tar = join_current_dir_redi(target);
 	if (ast_tar)
@@ -10,17 +35,17 @@ int asterisk_in_filename(char *target, t_cmd *cmd, t_token *tok)
 	else
 	{
 		g_herdoc_stop = true;
-		write (2, "wild_card error\n", 16);
+		write(2, "wild_card error\n", 16);
 	}
 	return (1);
 }
 
-char *get_single_file_or_null(char *patern)
+char	*get_single_file_or_null(char *patern)
 {
 	DIR				*dir_files;
 	struct dirent	*dir;
-	char *filename;
-	
+	char			*filename;
+
 	dir_files = opendir(".");
 	filename = NULL;
 	while (dir_files)
@@ -46,9 +71,38 @@ char *get_single_file_or_null(char *patern)
 
 char	*join_current_dir_redi(char *patern)
 {
-	char *new_target;
+	char	*new_target;
 
 	new_target = get_single_file_or_null(patern);
 	free(patern);
 	return (new_target);
+}
+
+bool	next_joined_word_is_pattern(t_lexer *lexer)
+{
+	t_token	next_tok;
+	t_token	tok;
+	t_lexer	lexer_back;
+	size_t	i;
+
+	lexer_back = *lexer;
+	while (1)
+	{
+		tok = lexer_next_token(lexer);
+		if (tok.type == TOK_SINGLE || tok.type == TOK_DOUBLE)
+		{
+			i = 0;
+			while (i < tok.len)
+			{
+				if (tok.literal[i] == '*')
+					return (*lexer = lexer_back, false);
+				i++;
+			}
+		}
+		next_tok = lexer_peek_next_token(lexer);
+		if (wds(&tok) == 0 || next_tok.space == true)
+			break ;
+	}
+	*lexer = lexer_back;
+	return (true);
 }
